addingTS and moving folders to lib
