+++ lib/data/CoursesDataSheet.js
@@ -2,8 +2,6 @@ import Iconinteract from "../svgs/interact.svg";
 import Iconvideo from "../svgs/video.svg";
 import IcontxtLocucion from "../svgs/txtLocucion.svg";
 import Iconautoeval from "../svgs/autoeval.svg";
-import Iconcuestionario from "../svgs/cuestionario.svg";
-
 
 export const DataSheet = {
   library: { name: "Recursos Humanos", skuPrefix: "R" },
