+++ lib/data/DataCourseLibrarie.js
@@ -1,8 +1,3 @@
-import Iconinteract from "../svgs/interact.svg";
-import Iconvideo from "../svgs/video.svg";
-import IcontxtLocucion from "../svgs/txtLocucion.svg";
-import Iconautoeval from "../svgs/autoeval.svg";
-import Iconcuestionario from "../svgs/cuestionario.svg";
 import ThumbnailHBEA from "../../public/thumbnail/thumbnail_R_HB_EA.jpg";
 import ThumbnailHBME from "../../public/thumbnail/thumbnail_R_HB_ME.jpg";
 import ThumbnailHD01 from "../../public/thumbnail/thumbnail_R_HD_01.jpg";
