+++ lib/helpers/library.js
