+++ package.json
@@ -37,10 +37,13 @@
   },
   "devDependencies": {
     "@svgr/webpack": "^6.3.1",
+    "@types/node": "^18.11.14",
+    "@types/react": "^18.0.26",
     "autoprefixer": "^10.4.11",
     "eslint": "8.23.1",
     "eslint-config-next": "12.3.1",
     "postcss": "^8.4.16",
-    "tailwindcss": "^3.1.8"
+    "tailwindcss": "^3.1.8",
+    "typescript": "^4.9.4"
   }
 }
