+++ pages/gracias.js
@@ -1,7 +1,7 @@
 import ImgMessagePhone from "../public/sendMessage.jpg";
 import Image from "next/image";
 import IconSend from "../public/icons/iconSent.svg";
-import HelperHead from "../components/helpers/HelperHead";
+import HelperHead from "../lib/helpers/HelperHead";
 import { DataHeadGracias } from "../components/data/DataHeader";
 
 function Gracias() {
